Move ModuleLocator and related interfaces to @glimmer/interfaces
Because ModuleLocator was contained in the @glimmer/bundle-compiler package, runtime libraries had to include this package even if they were only working with data produced by build tools.

Concretely, the current factoring was causing @glimmer/application to have a dependency on @glimmer/bundle-compiler for a single interface.

This commit moves ModuleLocator, AnnotatedModuleLocator and TemplateLocator to @glimmer/interfaces.
