+++ packages/@glimmer/bundle-compiler/index.ts
@@ -5,12 +5,7 @@ export {
   BundleCompilationResult
 } from './lib/bundle-compiler';
 
-export {
-  ModuleLocator,
-  AnnotatedModuleLocator,
-  TemplateLocator,
-  ModuleLocatorMap
-} from './lib/module-locators';
+export { default as ModuleLocatorMap } from './lib/module-locator-map';
 export { default as CompilerDelegate } from './lib/compiler-delegate';
 export { default as CompilerResolver } from './lib/compiler-resolver';
 export { default as DebugConstants } from './lib/debug-constants';
