+++ packages/@glimmer/test-helpers/lib/environment/components/emberish-curly.ts
@@ -1,5 +1,5 @@
 import { WrappedBuilder } from '@glimmer/opcode-compiler';
-import { Option, Opaque, ProgramSymbolTable, ComponentCapabilities } from '@glimmer/interfaces';
+import { Option, Opaque, ProgramSymbolTable, ComponentCapabilities, ModuleLocator } from '@glimmer/interfaces';
 import GlimmerObject from '@glimmer/object';
 import { Tag, combine, PathReference, TagWrapper, DirtyableTag } from '@glimmer/reference';
 import { EMPTY_ARRAY, assign, Destroyable, expect } from '@glimmer/util';
@@ -10,7 +10,6 @@ import { Attrs, createTemplate, AttrsDiff } from '../shared';
 import LazyRuntimeResolver from '../modes/lazy/runtime-resolver';
 import EagerRuntimeResolver from '../modes/eager/runtime-resolver';
 import { TestComponentDefinitionState, TemplateMeta } from '../components';
-import { ModuleLocator } from '@glimmer/bundle-compiler';
 
 export class EmberishCurlyComponent extends GlimmerObject {
   public static positionalParams: string[] | string = [];
